Create CleanupRequest middleware to really cleanup requests for REST2
Previously we added the cleanup functionality to to_psgi_app(initially
created as an alias of "to_app"), which never gets called. The call of
response_cb seems not quite right either as it doesn't expect a Plack
app as the first argument.

This commit wraps the functionality into a middleware and calls the more
complete version(RT::Interface::Web::Handler::CleanupRequest).